fix: Change type int to long in file_size fields

+++ src/library/Telegram.BotAPI/AvailableTypes/File.cs
@@ -25,7 +25,7 @@ public class File
     /// Optional. File size in bytes. It can be bigger than 2^31 and some programming languages may have difficulty/silent defects in interpreting it. But it has at most 52 significant bits, so a signed 64-bit integer or double-precision float type are safe for storing this value.
     /// </summary>
     [JsonPropertyName(PropertyNames.FileSize)]
-    public int? FileSize { get; set; }
+    public long? FileSize { get; set; }
 
     /// <summary>
     /// Optional. File path. Use <em>https://api.telegram.org/file/bot&lt;token&gt;/&lt;file_path&gt;</em> to get the file.
