+++ src/library/Telegram.BotAPI/AvailableTypes/PhotoSize.cs
@@ -37,5 +37,5 @@ public class PhotoSize
     /// Optional. File size in bytes
     /// </summary>
     [JsonPropertyName(PropertyNames.FileSize)]
-    public int? FileSize { get; set; }
+    public long? FileSize { get; set; }
 }
